Move plot_sensitivity to rmgpy.tools.plot
To avoid some cyclic imports, and it seems like a better location
